requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
defusedxml==0.7.1
exceptiongroup==1.2.2
fastapi==0.114.1
h11==0.14.0
idna==3.8
odfpy==1.4.1
pydantic==2.9.1
pydantic_core==2.23.3
python-multipart==0.0.9
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
